requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
Django==3.2.9
django-redis==5.0.0
djangorestframework==3.12.4
dnspython==2.1.0
email-validator==1.1.3
idna==3.3
Jinja2==3.0.2
MarkupSafe==2.0.1
pydantic==1.8.2
pytz==2021.3
questions==0.7.0a4
redis==3.5.3
rejson==0.5.4
requests==2.26.0
simpleeval==0.9.11
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.7
